fixes
so if I want to be able to have the main method in each java file, I have to have a class name that doesn't start with a number. and also VSCode doesn't like the package names..
So. I'm moving things out of their folders and just renaming every problem as SolutionX where X is the problem number.